Did work to generate initial requests for web servers as well as the web servers themselves.

+++ Request.cpp
@@ -2,7 +2,7 @@
 
 #include "Request.h"
 
-Request::Request(const std::string &requestIp, const std::string &destinationIp, int processingTime)
+Request::Request(const std::string &requestIp, const std::string &destinationIp, float processingTime)
     : requestIp(requestIp), destinationIp(destinationIp), processingTime(processingTime) {}
 
 std::string Request::getRequestIp() const
